Tweaks, reciepts, test build, version .6 alpha

+++ Voco/Voco/API/VCObjectRequestOperation.m
@@ -48,7 +48,7 @@
                 if(apiError != nil){
                     [UIAlertView showWithTitle:@"Error" message:apiError.error cancelButtonTitle:@"Oh, ok" otherButtonTitles:nil tapBlock:nil];
                 } else {
-                    NSLog(@"Error %i", statusCode);
+                    NSLog(@"Error %li", (long)statusCode);
                     [UIAlertView showWithTitle:@"Error" message:@"Unspecified Error" cancelButtonTitle:@"Ok, I'll try that again I guess" otherButtonTitles:nil tapBlock:nil];
                 }
             }
